Code has been refactored
