+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "didibudget-frontend",
-  "version": "0.7.3",
+  "version": "0.7.4",
   "description": "didibudget-frontend",
   "private": true,
   "dependencies": {
