+++ src/pages/Users/UserAdministration.js
@@ -1,8 +1,8 @@
 import React, { Fragment } from 'react'
 
-import { PageTitle } from '../components/PageTitle'
+import { PageTitle } from '../../components/PageTitle'
 
-import { GetListOfUsers } from '../components/GetListOfUsers'
+import { GetListOfUsers } from '../../components/Users/GetListOfUsers'
 
 const UserAdministration = () => {
 	return (
